+++ src/Header/AcceptLanguage.php
@@ -3,26 +3,43 @@ declare(strict_types = 1);
 
 namespace Innmind\Http\Header;
 
-use Innmind\Http\Header;
+use Innmind\Http\{
+    Header,
+    Header\Accept\Language,
+};
+use Innmind\Immutable\Sequence;
 
 /**
  * @psalm-immutable
  */
 final class AcceptLanguage implements Custom
 {
-    private Header $header;
+    /**
+     * @param Sequence<Language> $languages
+     */
+    private function __construct(
+        private Sequence $languages,
+    ) {
+    }
 
     /**
+     * @psalm-pure
      * @no-named-arguments
      */
-    public function __construct(AcceptLanguageValue ...$values)
+    public static function of(Language ...$languages): self
     {
-        $this->header = new Header('Accept-Language', ...$values);
+        return new self(Sequence::of(...$languages));
     }
 
     #[\Override]
     public function normalize(): Header
     {
-        return $this->header;
+        return new Header(
+            'Accept-Language',
+            ...$this
+                ->languages
+                ->map(static fn($value) => new Value\Value($value->toString()))
+                ->toList(),
+        );
     }
 }