+++ tests/Header/AcceptLanguageTest.php
@@ -6,8 +6,8 @@ namespace Tests\Innmind\Http\Header;
 use Innmind\Http\{
     Header\AcceptLanguage,
     Header,
-    Header\AcceptLanguageValue,
-    Header\Parameter\Quality
+    Header\Accept\Language,
+    Header\Parameter\Quality,
 };
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
@@ -15,9 +15,12 @@ class AcceptLanguageTest extends TestCase
 {
     public function testInterface()
     {
-        $h = new AcceptLanguage(
-            $v = new AcceptLanguageValue('fr', new Quality(0.8)),
-        );
+        $h = Language::maybe('fr', new Quality(0.8))
+            ->map(AcceptLanguage::of(...))
+            ->match(
+                static fn($header) => $header,
+                static fn() => null,
+            );
 
         $this->assertInstanceOf(Header\Custom::class, $h);
         $this->assertSame('Accept-Language: fr;q=0.8', $h->normalize()->toString());
@@ -25,6 +28,6 @@ class AcceptLanguageTest extends TestCase
 
     public function testWithoutValues()
     {
-        $this->assertSame('Accept-Language: ', (new AcceptLanguage)->normalize()->toString());
+        $this->assertSame('Accept-Language: ', AcceptLanguage::of()->normalize()->toString());
     }
 }